+++ ksadk/server/static/assets/architecture-7EHR7CIX-CAFeU8Os.js
@@ -1,1 +1,0 @@
-import{x as e}from"./mermaid-parser.core-D7S0E4RN.js";export{e as createArchitectureServices};