+++ ksadk/server/static/assets/architecture-7EHR7CIX-ld802pvW.js
@@ -1,0 +1,1 @@
+import{x as e}from"./mermaid-parser.core-Cc12WCm1.js";export{e as createArchitectureServices};